Modify pom.xml to publish the component to Maven Central
Motivation:

The fix to netty HTTP3 is merged into the Netty HTTP3 repository. But the next version
that would include the fix has not been out yet.  In the mean time, this package needs
to be available in public.

Modifications:

Modified pom.xml to make this package available as io.github.naokiiwakami:netty-incubator-codec-http3

Result:

This package has been released via Maven Central.